feat(Console): add console verify endpoint
The console verify endpoint is needed by the ws server,
when the ws server receive a connection from a console overlay,
the ws server need to validate this connection using console_id
and console_token fields provided by the overlay.
The ws server send a request to `POST /console/verify` to verify
the validity of these and then accept or reject the client.